+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This files contains the list of changes between different versions of Upload Ext
 * [FEATURE] Built-in Extensions Manager.
 * [FEATURE] Loading status in JavaScript.
 * [FEATURE] Navigation and menu bars.
+* [FEATURE] Downloads of properly packaged extensions.
 
 ### Improvements